Separate base test classes
