+++ tests/agents_tests/test_double_dqn.py
@@ -7,20 +7,11 @@ from future import standard_library
 standard_library.install_aliases()
 
 from chainerrl.agents.double_dqn import DoubleDQN
-from test_dqn_like import _TestDQNOnContinuousABC
-from test_dqn_like import _TestDQNOnDiscreteABC
-from test_dqn_like import _TestDQNOnDiscretePOABC
+
+import basetest_dqn_like
 
 
-class TestDoubleDQNOnDiscreteABC(_TestDQNOnDiscreteABC):
-
-    def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
-        return DoubleDQN(
-            q_func, opt, rbuf, gpu=gpu, gamma=0.9, explorer=explorer,
-            replay_start_size=100, target_update_interval=100)
-
-
-class TestDoubleDQNOnContinuousABC(_TestDQNOnContinuousABC):
+class TestDoubleDQNOnDiscreteABC(basetest_dqn_like._TestDQNOnDiscreteABC):
 
     def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
         return DoubleDQN(
@@ -28,7 +19,15 @@ class TestDoubleDQNOnContinuousABC(_TestDQNOnContinuousABC):
             replay_start_size=100, target_update_interval=100)
 
 
-class TestDoubleDQNOnDiscretePOABC(_TestDQNOnDiscretePOABC):
+class TestDoubleDQNOnContinuousABC(basetest_dqn_like._TestDQNOnContinuousABC):
+
+    def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
+        return DoubleDQN(
+            q_func, opt, rbuf, gpu=gpu, gamma=0.9, explorer=explorer,
+            replay_start_size=100, target_update_interval=100)
+
+
+class TestDoubleDQNOnDiscretePOABC(basetest_dqn_like._TestDQNOnDiscretePOABC):
 
     def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
         return DoubleDQN(
