+++ tests/agents_tests/test_double_pal.py
@@ -7,20 +7,11 @@ from builtins import *  # NOQA
 standard_library.install_aliases()
 
 from chainerrl.agents.double_pal import DoublePAL
-from test_dqn_like import _TestDQNOnContinuousABC
-from test_dqn_like import _TestDQNOnDiscreteABC
-from test_dqn_like import _TestDQNOnDiscretePOABC
+
+import basetest_dqn_like
 
 
-class TestDoublePALOnDiscreteABC(_TestDQNOnDiscreteABC):
-
-    def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
-        return DoublePAL(
-            q_func, opt, rbuf, gpu=gpu, gamma=0.9, explorer=explorer,
-            replay_start_size=100, target_update_interval=100)
-
-
-class TestDoublePALOnContinuousABC(_TestDQNOnContinuousABC):
+class TestDoublePALOnDiscreteABC(basetest_dqn_like._TestDQNOnDiscreteABC):
 
     def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
         return DoublePAL(
@@ -28,7 +19,15 @@ class TestDoublePALOnContinuousABC(_TestDQNOnContinuousABC):
             replay_start_size=100, target_update_interval=100)
 
 
-class TestDoublePALOnDiscretePOABC(_TestDQNOnDiscretePOABC):
+class TestDoublePALOnContinuousABC(basetest_dqn_like._TestDQNOnContinuousABC):
+
+    def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
+        return DoublePAL(
+            q_func, opt, rbuf, gpu=gpu, gamma=0.9, explorer=explorer,
+            replay_start_size=100, target_update_interval=100)
+
+
+class TestDoublePALOnDiscretePOABC(basetest_dqn_like._TestDQNOnDiscretePOABC):
 
     def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
         return DoublePAL(
