+++ tests/agents_tests/test_dqn.py
@@ -6,21 +6,19 @@ from future import standard_library
 from builtins import *  # NOQA
 standard_library.install_aliases()
 
+import basetest_dqn_like as base
 import chainerrl
 from chainerrl.agents.dqn import DQN
-from test_dqn_like import _TestDQNOnContinuousABC
-from test_dqn_like import _TestDQNOnDiscreteABC
-from test_dqn_like import _TestDQNOnDiscretePOABC
 
 
-class TestDQNOnDiscreteABC(_TestDQNOnDiscreteABC):
+class TestDQNOnDiscreteABC(base._TestDQNOnDiscreteABC):
 
     def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
         return DQN(q_func, opt, rbuf, gpu=gpu, gamma=0.9, explorer=explorer,
                    replay_start_size=100, target_update_interval=100)
 
 
-class TestDQNOnDiscreteABCBoltzmann(_TestDQNOnDiscreteABC):
+class TestDQNOnDiscreteABCBoltzmann(base._TestDQNOnDiscreteABC):
 
     def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
         explorer = chainerrl.explorers.Boltzmann()
@@ -28,14 +26,14 @@ class TestDQNOnDiscreteABCBoltzmann(_TestDQNOnDiscreteABC):
                    replay_start_size=100, target_update_interval=100)
 
 
-class TestDQNOnContinuousABC(_TestDQNOnContinuousABC):
+class TestDQNOnContinuousABC(base._TestDQNOnContinuousABC):
 
     def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
         return DQN(q_func, opt, rbuf, gpu=gpu, gamma=0.9, explorer=explorer,
                    replay_start_size=100, target_update_interval=100)
 
 
-class TestDQNOnDiscretePOABC(_TestDQNOnDiscretePOABC):
+class TestDQNOnDiscretePOABC(base._TestDQNOnDiscretePOABC):
 
     def make_dqn_agent(self, env, q_func, opt, explorer, rbuf, gpu):
         return DQN(q_func, opt, rbuf, gpu=gpu, gamma=0.9, explorer=explorer,
